End of section 5. Add Identity service and connect it to ActionService

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
 using MassTransit.Testing;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -51,14 +52,14 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(acution);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
 
-        //TODO: add current user as seller
-        auction.Seller = "test";
-
+        auction.Seller = User.Identity.Name;
+        
         _context.Auctions.Add(auction);
 
         var newAction = _mapper.Map<AuctionDto>(auction);
@@ -73,6 +74,7 @@ public class AuctionsController : ControllerBase
             new {auction.Id}, newAction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -81,7 +83,7 @@ public class AuctionsController : ControllerBase
 
         if (auction is null) return NotFound();
 
-        // TODO: check seller == username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -98,6 +100,7 @@ public class AuctionsController : ControllerBase
         return BadRequest("Problem saving changes");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -105,7 +108,7 @@ public class AuctionsController : ControllerBase
 
         if (auction is null) return NotFound();
 
-        // TODO: check seller == username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
